fix: remove redundant header components and unify the UI/UX
