+++ packages/core/src/App/Components/Layout/Header/index.js
@@ -2,6 +2,6 @@ import ToggleNotifications from './toggle-notifications.jsx';
 import PlatformSwitcher from './platform-switcher.jsx';
 
 export * from './menu-links.jsx';
-export * from './account-actions.jsx';
+export * from './account-actions.tsx';
 export * from './wallets/account-actions-wallets';
 export { ToggleNotifications, PlatformSwitcher };
