+++ packages/core/src/App/Containers/Layout/header/header-legacy.tsx
@@ -1,12 +1,10 @@
 import React from 'react';
-import { useHistory } from 'react-router-dom';
+import { useHistory, useLocation } from 'react-router-dom';
 import classNames from 'classnames';
-
 import { getDecimalPlaces, platforms, routes } from '@deriv/shared';
 import { observer, useStore } from '@deriv/stores';
 import { useDevice } from '@deriv-com/ui';
-
-import { MenuLinks, PlatformSwitcher } from 'App/Components/Layout/Header';
+import { MenuLinks } from 'App/Components/Layout/Header';
 import { AccountsInfoLoader } from 'App/Components/Layout/Header/Components/Preloader';
 import ToggleMenuDrawer from 'App/Components/Layout/Header/toggle-menu-drawer.jsx';
 import ToggleMenuDrawerAccountsOS from 'App/Components/Layout/Header/toggle-menu-drawer-accounts-os.jsx';
@@ -14,13 +12,12 @@ import platform_config from 'App/Constants/platform-config';
 import NewVersionNotification from 'App/Containers/new-version-notification.jsx';
 import RealAccountSignup from 'App/Containers/RealAccountSignup';
 import SetAccountCurrencyModal from 'App/Containers/SetAccountCurrencyModal';
-
 import DerivShortLogo from './deriv-short-logo';
 import HeaderAccountActions from './header-account-actions';
 import TradersHubHomeButton from './traders-hub-home-button';
 
-const DTraderHeader = observer(() => {
-    const { client, common, ui, notifications, traders_hub } = useStore();
+const HeaderLegacy = observer(() => {
+    const { client, common, ui, notifications } = useStore();
     const {
         currency,
         has_any_real_account,
@@ -33,14 +30,21 @@ const DTraderHeader = observer(() => {
         is_virtual,
         is_switching,
     } = client;
-    const { app_routing_history, platform, current_language, is_from_tradershub_os } = common;
+    const { platform, is_from_tradershub_os } = common;
     const { header_extension, is_app_disabled, is_route_modal_on, toggleReadyToDepositModal, is_real_acc_signup_on } =
         ui;
     const { addNotificationMessage, client_notifications, removeNotificationMessage } = notifications;
-    const { setTogglePlatformType } = traders_hub;
+
+    const { isDesktop } = useDevice();
 
     const history = useHistory();
-    const { isDesktop, isMobile } = useDevice();
+    const { pathname } = useLocation();
+
+    const traders_hub_routes =
+        [routes.traders_hub].includes(pathname) ||
+        [routes.account, routes.settings, routes.wallets_compare_accounts, routes.compare_cfds].some(route =>
+            pathname.startsWith(route)
+        );
 
     const addUpdateNotification = () => addNotificationMessage(client_notifications?.new_version_available);
     const removeUpdateNotification = React.useCallback(
@@ -86,7 +90,13 @@ const DTraderHeader = observer(() => {
         >
             <div className='header__menu-items'>
                 <div className='header__menu-left'>
-                    {!isDesktop ? (
+                    {isDesktop ? (
+                        <React.Fragment>
+                            <DerivShortLogo />
+                            <div className='header__divider' />
+                            <TradersHubHomeButton />
+                        </React.Fragment>
+                    ) : (
                         <React.Fragment>
                             {is_from_tradershub_os ? (
                                 <ToggleMenuDrawerAccountsOS
@@ -100,21 +110,10 @@ const DTraderHeader = observer(() => {
                                     )}
                                 </>
                             )}
-                        </React.Fragment>
-                    ) : (
-                        <React.Fragment>
                             <DerivShortLogo />
-                            <div className='header__divider' />
-                            <TradersHubHomeButton />
-                            <PlatformSwitcher
-                                app_routing_history={app_routing_history}
-                                platform_config={filterPlatformsForClients(platform_config)}
-                                setTogglePlatformType={setTogglePlatformType}
-                                current_language={current_language}
-                            />
                         </React.Fragment>
                     )}
-                    <MenuLinks />
+                    <MenuLinks {...{ is_traders_hub_routes: traders_hub_routes }} />
                 </div>
 
                 <div
@@ -122,11 +121,6 @@ const DTraderHeader = observer(() => {
                         'header__menu-right--hidden': !isDesktop && is_logging_in,
                     })}
                 >
-                    {isDesktop && (
-                        <div className='header__menu--dtrader--separator--account'>
-                            <div className='header__menu--dtrader--separator' />
-                        </div>
-                    )}
                     {(is_logging_in || is_single_logging_in || is_switching) && (
                         <div
                             id='dt_core_header_acc-info-preloader'
@@ -138,7 +132,12 @@ const DTraderHeader = observer(() => {
                             <AccountsInfoLoader is_logged_in={is_logged_in} is_mobile={!isDesktop} speed={3} />
                         </div>
                     )}
-                    {!is_from_tradershub_os && <HeaderAccountActions onClickDeposit={handleClickCashier} />}
+                    {!is_from_tradershub_os && (
+                        <HeaderAccountActions
+                            onClickDeposit={handleClickCashier}
+                            is_traders_hub_routes={traders_hub_routes}
+                        />
+                    )}
                 </div>
             </div>
             {is_real_acc_signup_on && <RealAccountSignup />}
@@ -148,4 +147,4 @@ const DTraderHeader = observer(() => {
     );
 });
 
-export default DTraderHeader;
+export default HeaderLegacy;
