+++ packages/core/src/App/Containers/Layout/header/header-wallets.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
 import classNames from 'classnames';
-import { withRouter } from 'react-router-dom';
+import { withRouter, useLocation } from 'react-router-dom';
 import { useDevice } from '@deriv-com/ui';
 import { observer, useStore } from '@deriv/stores';
 import { routes, platforms } from '@deriv/shared';
-import { MenuLinks, PlatformSwitcher } from 'App/Components/Layout/Header';
+import { MenuLinks } from 'App/Components/Layout/Header';
 import { AccountActionsWallets } from 'App/Components/Layout/Header/wallets/account-actions-wallets';
 import platform_config from 'App/Constants/platform-config';
 import RealAccountSignup from 'App/Containers/RealAccountSignup';
@@ -13,15 +13,21 @@ import NewVersionNotification from 'App/Containers/new-version-notification.jsx'
 import ToggleMenuDrawer from 'App/Components/Layout/Header/toggle-menu-drawer.jsx';
 import AccountsInfoLoaderWallets from 'App/Components/Layout/Header/wallets/accounts-info-loader-wallets';
 import TradersHubHomeButton from './traders-hub-home-button';
+import DerivShortLogo from './deriv-short-logo';
 
 const MenuLeft = observer(() => {
-    const { client, common, ui, traders_hub } = useStore();
-    const { is_bot_allowed, is_logged_in, is_mt5_allowed, is_dxtrade_allowed } = client;
-    const { app_routing_history, current_language } = common;
-    const { header_extension } = ui;
-    const { setTogglePlatformType } = traders_hub;
-
+    const { client, ui } = useStore();
     const { isDesktop } = useDevice();
+    const { pathname } = useLocation();
+
+    const { is_bot_allowed, is_logged_in, is_mt5_allowed, is_dxtrade_allowed } = client;
+    const { header_extension } = ui;
+
+    const traders_hub_routes =
+        [routes.traders_hub].includes(pathname) ||
+        [routes.account, routes.wallets, routes.settings, routes.wallets_compare_accounts, routes.compare_cfds].some(
+            route => pathname.startsWith(route)
+        );
 
     const filterPlatformsForClients = (payload: typeof platform_config) =>
         payload.filter(config => {
@@ -41,14 +47,9 @@ const MenuLeft = observer(() => {
         <div className='header__menu-left'>
             {isDesktop ? (
                 <React.Fragment>
+                    <DerivShortLogo />
+                    <div className='header__divider' />
                     <TradersHubHomeButton />
-                    <div className='traders-hub-header__divider traders-hub-header__divider--wallets' />
-                    <PlatformSwitcher
-                        app_routing_history={app_routing_history}
-                        platform_config={filterPlatformsForClients(platform_config)}
-                        setTogglePlatformType={setTogglePlatformType}
-                        current_language={current_language}
-                    />
                 </React.Fragment>
             ) : (
                 <React.Fragment>
@@ -56,17 +57,25 @@ const MenuLeft = observer(() => {
                     {header_extension && is_logged_in && (
                         <div className='header__menu-left-extensions'>{header_extension}</div>
                     )}
+                    <DerivShortLogo />
                 </React.Fragment>
             )}
-            <MenuLinks />
+            <MenuLinks {...{ is_traders_hub_routes: traders_hub_routes }} />
         </div>
     );
 });
 
 const MenuRight = observer(() => {
+    const { pathname } = useLocation();
     const { client, ui } = useStore();
     const { is_logged_in, is_logging_in, is_single_logging_in, is_switching, accounts, loginid, is_crypto } = client;
     const { is_mobile } = ui;
+
+    const traders_hub_routes =
+        [routes.traders_hub].includes(pathname) ||
+        [routes.account, routes.wallets, routes.settings, routes.wallets_compare_accounts, routes.compare_cfds].some(
+            route => pathname.startsWith(route)
+        );
 
     const active_account = accounts?.[loginid ?? ''];
     const currency = active_account?.currency ?? '';
@@ -85,14 +94,14 @@ const MenuRight = observer(() => {
                 </div>
             ) : (
                 <div id={'dt_core_header_acc-info-container'} className='acc-info__container'>
-                    <AccountActionsWallets />
+                    <AccountActionsWallets is_traders_hub_routes={traders_hub_routes} />
                 </div>
             )}
         </div>
     );
 });
 
-const DTraderHeaderWallets = observer(() => {
+const HeaderWallets = observer(() => {
     const { common, ui, notifications } = useStore();
     const { platform } = common;
     const { is_app_disabled, is_route_modal_on } = ui;
@@ -127,4 +136,4 @@ const DTraderHeaderWallets = observer(() => {
     );
 });
 
-export default withRouter(DTraderHeaderWallets);
+export default withRouter(HeaderWallets);
