+++ packages/core/src/App/Containers/Layout/header/header.tsx
@@ -3,62 +3,26 @@ import { useLocation } from 'react-router-dom';
 import { useReadLocalStorage } from 'usehooks-ts';
 import { makeLazyLoader, moduleLoader, routes } from '@deriv/shared';
 import { observer, useStore } from '@deriv/stores';
-import { useDevice } from '@deriv-com/ui';
 import classNames from 'classnames';
 
 const HeaderFallback = () => {
     return <div className={classNames('header')} />;
 };
 
-const DefaultHeader = makeLazyLoader(
-    () => moduleLoader(() => import(/* webpackChunkName: "default-header" */ './default-header')),
+const HeaderLegacy = makeLazyLoader(
+    () => moduleLoader(() => import(/* webpackChunkName: "dtrader-header" */ './header-legacy')),
     () => <HeaderFallback />
 )();
 
-const DefaultHeaderWallets = makeLazyLoader(
-    () => moduleLoader(() => import(/* webpackChunkName: "default-header-wallets" */ './default-header-wallets')),
-    () => <HeaderFallback />
-)();
-
-const DTraderHeader = makeLazyLoader(
-    () => moduleLoader(() => import(/* webpackChunkName: "dtrader-header" */ './dtrader-header')),
-    () => <HeaderFallback />
-)();
-
-const DTraderHeaderWallets = makeLazyLoader(
-    () => moduleLoader(() => import(/* webpackChunkName: "dtrader-header-wallets" */ './dtrader-header-wallets')),
-    () => <HeaderFallback />
-)();
-
-const TradersHubHeader = makeLazyLoader(
-    () => moduleLoader(() => import(/* webpackChunkName: "traders-hub-header" */ './traders-hub-header')),
-    () => <HeaderFallback />
-)();
-
-const TradersHubHeaderWallets = makeLazyLoader(
-    () =>
-        moduleLoader(() => import(/* webpackChunkName: "traders-hub-header-wallets" */ './traders-hub-header-wallets')),
+const HeaderWallets = makeLazyLoader(
+    () => moduleLoader(() => import(/* webpackChunkName: "dtrader-header-wallets" */ './header-wallets')),
     () => <HeaderFallback />
 )();
 
 const Header = observer(() => {
-    const { client, common } = useStore();
-    const { accounts, has_wallet, is_logged_in, setAccounts, loginid, switchAccount, is_single_logging_in } = client;
-    const { is_from_tradershub_os } = common;
+    const { client } = useStore();
+    const { accounts, has_wallet, is_logged_in, setAccounts, loginid, switchAccount } = client;
     const { pathname } = useLocation();
-
-    const is_wallets_cashier_route = pathname.includes(routes.wallets);
-
-    const traders_hub_routes =
-        [
-            routes.traders_hub,
-            routes.account,
-            routes.cashier,
-            routes.wallets_compare_accounts,
-            routes.compare_cfds,
-        ].includes(pathname) ||
-        pathname.startsWith(routes.compare_cfds) ||
-        is_wallets_cashier_route;
 
     const client_accounts = useReadLocalStorage('client.accounts');
     React.useEffect(() => {
@@ -72,31 +36,9 @@ const Header = observer(() => {
             }
         }
     }, [accounts, client_accounts, has_wallet, is_logged_in, loginid, setAccounts, switchAccount]);
-    if (is_logged_in && !is_single_logging_in) {
-        let result;
-        switch (true) {
-            case pathname === routes.onboarding:
-                result = null;
-                break;
-            case traders_hub_routes:
-                result = has_wallet ? <TradersHubHeaderWallets /> : <TradersHubHeader />;
-                break;
-            case pathname.includes(routes.account):
-                if (is_from_tradershub_os) {
-                    result = <DTraderHeader />;
-                } else {
-                    result = has_wallet ? <DTraderHeaderWallets /> : <DTraderHeader />;
-                }
-                break;
-            default:
-                result = has_wallet ? <DTraderHeaderWallets /> : <DTraderHeader />;
-                break;
-        }
-        return result;
-    } else if (pathname === routes.onboarding) {
-        return null;
-    }
-    return has_wallet ? <DefaultHeaderWallets /> : <DefaultHeader />;
+
+    if (pathname === routes.onboarding) return null;
+    return has_wallet ? <HeaderWallets /> : <HeaderLegacy />;
 });
 
 export default Header;
