+++ packages/core/src/App/Containers/Layout/header/traders-hub-onboarding.tsx
@@ -27,22 +27,20 @@ const TradersHubOnboarding = observer(() => {
     );
 
     return (
-        <div data-testid='dt_traders_hub_onboarding'>
-            <div className='traders-hub-header__tradershub--onboarding--logo'>
-                {isTabletOs ? (
-                    onboardingIcon
-                ) : (
-                    <Popover
-                        classNameBubble='account-settings-toggle__tooltip'
-                        alignment='bottom'
-                        message={!is_mobile && <Localize i18n_default_text='View tutorial' />}
-                        should_disable_pointer_events
-                        zIndex='9999'
-                    >
-                        {onboardingIcon}
-                    </Popover>
-                )}
-            </div>
+        <div data-testid='dt_traders_hub_onboarding' className='traders-hub-onboarding__toggle'>
+            {isTabletOs ? (
+                onboardingIcon
+            ) : (
+                <Popover
+                    classNameBubble='account-settings-toggle__tooltip'
+                    alignment='bottom'
+                    message={!is_mobile && <Localize i18n_default_text='View tutorial' />}
+                    should_disable_pointer_events
+                    zIndex='9999'
+                >
+                    {onboardingIcon}
+                </Popover>
+            )}
         </div>
     );
 });
