+++ packages/core/src/sass/app/_common/components/account-settings-toggle.scss
@@ -1,4 +1,5 @@
 .account-settings-toggle {
+    margin-right: 1.2rem;
     svg {
         height: 20px;
         width: 20px;
