+++ packages/core/src/sass/app/_common/components/account-switcher.scss
@@ -6,7 +6,7 @@
     flex-direction: row;
     height: 100%;
     justify-content: center;
-    padding: 0 1.6rem 0 1.7rem;
+    padding: 0 1.6rem;
 
     &__preloader {
         position: absolute;
@@ -88,9 +88,11 @@
     }
 
     &__container {
+        display: flex;
         align-items: center;
         -webkit-box-align: center;
-        display: flex;
+        padding-right: 0.8rem;
+        gap: 1.6rem;
     }
     &__wrapper {
         align-items: center;
@@ -99,7 +101,6 @@
         height: 100%;
         justify-content: center;
         position: relative;
-        margin-right: 0.8rem;
         user-select: none;
         -webkit-touch-callout: none;
         -webkit-tap-highlight-color: transparent;
@@ -114,9 +115,7 @@
     }
     &__separator {
         border-right: 1px solid var(--general-section-1);
-        width: 2rem;
         height: 3.2rem;
-        margin-right: -0.1rem;
     }
     &__balance {
         @include typeface(--title-center-bold-green, default);
@@ -181,11 +180,6 @@
         }
     }
     &__button {
-        margin: 0 1.6em 0 0 !important;
-
-        &:not(:last-child) {
-            margin-right: 1em;
-        }
         @include mobile-screen {
             height: 2.8rem !important;
         }
@@ -194,10 +188,6 @@
     &__wallets {
         padding: 1.2rem 1.6rem;
         gap: 0.8rem;
-
-        &-notification-icon {
-            margin-right: 1rem;
-        }
 
         &-container {
             align-items: center;
