+++ packages/core/src/sass/app/_common/layout/header.scss
@@ -178,7 +178,7 @@
 
         &-logo {
             align-self: center;
-            margin: 0 1.2rem;
+            margin: 0 1.6rem;
             cursor: pointer;
         }
     }
@@ -393,7 +393,6 @@
     &__divider {
         width: 0.1rem;
         height: 3rem;
-        margin-left: 0.8rem;
         background: var(--general-section-2);
 
         @include mobile {
